Fixed a dialogue flow bug.
