+++ src/dev/ipsych0/myrinnia/entities/statics/Campfire.java
@@ -143,4 +143,9 @@ public class Campfire extends StaticEntity {
         Handler.get().getWorld().getEntityManager().addEntity(new Campfire(xSpawn, ySpawn));
     }
 
+    @Override
+    protected void updateDialogue() {
+
+    }
+
 }
