+++ src/dev/ipsych0/myrinnia/entities/statics/Tree.java
@@ -128,4 +128,9 @@ public class Tree extends StaticEntity {
         Handler.get().getWorld().getEntityManager().addEntity(new Tree(xSpawn, ySpawn));
     }
 
+    @Override
+    protected void updateDialogue() {
+
+    }
+
 }
