+++ src/dev/ipsych0/myrinnia/entities/statics/WaterToBridgePart.java
@@ -128,4 +128,9 @@ public class WaterToBridgePart extends StaticEntity {
 
     }
 
+    @Override
+    protected void updateDialogue() {
+
+    }
+
 }
